+++ libs/community/tests/integration_tests/vectorstores/qdrant/async_api/test_max_marginal_relevance.py
@@ -10,6 +10,7 @@ from tests.integration_tests.vectorstores.fake_embeddings import (
 from tests.integration_tests.vectorstores.qdrant.async_api.fixtures import (
     qdrant_locations,
 )
+from tests.integration_tests.vectorstores.qdrant.common import assert_documents_equals
 
 
 @pytest.mark.parametrize("batch_size", [1, 64])
@@ -41,7 +42,10 @@ async def test_qdrant_max_marginal_relevance_search(
     output = await docsearch.amax_marginal_relevance_search(
         "foo", k=2, fetch_k=3, lambda_mult=0.0
     )
-    assert output == [
-        Document(page_content="foo", metadata={"page": 0}),
-        Document(page_content="baz", metadata={"page": 2}),
-    ]
+    assert_documents_equals(
+        output,
+        [
+            Document(page_content="foo", metadata={"page": 0}),
+            Document(page_content="baz", metadata={"page": 2}),
+        ],
+    )
